# release-23.2: sql/stats: improve query to ensure all tables